Feature/role based access

+++ src/api-gateway/src/company/company.controller.ts
@@ -55,6 +55,7 @@ import { JwtService } from '@nestjs/jwt';
 import { BooleanResponseDto } from '../shared/dtos/api-response.dto';
 import { FileInterceptor } from '@nestjs/platform-express';
 import { JobStatusAllResponseDto } from '../job/dto/job-responses.dto';
+import { EmployeeService } from '../employee/employee.service';
 
 const className = 'Company';
 
@@ -64,6 +65,7 @@ export class CompanyController {
   constructor(
     private readonly companyService: CompanyService,
     private readonly jwtService: JwtService,
+    private readonly employeeService: EmployeeService,
   ) {}
   validateObjectId(id: string | Types.ObjectId): boolean {
     if (!mongoose.Types.ObjectId.isValid(id)) {
@@ -104,16 +106,28 @@ export class CompanyController {
   @ApiBody({ type: AddUserToCompanyDto })
   @ApiOkResponse({ type: BooleanResponseDto })
   @Post('/add')
-  async addEmployee(@Body() addUserDto: AddUserToCompanyDto) {
-    const arr = [addUserDto.adminId, addUserDto.currentCompany];
-    if (addUserDto.roleId) arr.push(addUserDto.roleId);
+  async addEmployee(
+    @Body()
+    body: {
+      currentEmployeeId: Types.ObjectId;
+      addUserDto: AddUserToCompanyDto;
+    },
+  ) {
+    // const currentEmployee = await this.employeeService.findById(
+    //   body.currentEmployeeId,
+    // );
+    // if (currentEmployee.role.permissionSuite.includes('add new employees')) {
+    //TODO: Figure out if this enpoint needs role based access
+    const arr = [body.addUserDto.adminId, body.addUserDto.currentCompany];
+    if (body.addUserDto.roleId) arr.push(body.addUserDto.roleId);
     validateObjectIds(arr);
 
     try {
-      return { data: await this.companyService.addEmployee(addUserDto) };
+      return { data: await this.companyService.addEmployee(body.addUserDto) };
     } catch (Error) {
       throw new HttpException('Internal server error', HttpStatus.CONFLICT);
     }
+    // }
   }
 
   @UseGuards(AuthGuard) //It may be accessed by external users
@@ -161,12 +175,53 @@ export class CompanyController {
     description: `The _id attribute of the ${className}`,
   })
   @Get('/all/employees/:cid')
-  async getAllEmployeesInCompany(@Param('cid') cid: string) {
+  async getAllEmployeesInCompany(
+    @Param('cid') cid: string,
+    // @Body() body: { currentEmployeeId: Types.ObjectId },
+  ) {
+    // console.log('In getAllEmployeesInCompany');
+    // console.log('body.currentEmployeeId: ', body.currentEmployeeId);
+    // const currentEmployee = await this.employeeService.findById(
+    //   body.currentEmployeeId,
+    // );
+    // console.log('Current Employee:', currentEmployee);
+    // if (currentEmployee.role.permissionSuite.includes('view all employees')) {
+    //   validateObjectId(cid);
+    //   const objId = new Types.ObjectId(cid);
+    //   try {
+    //     return { data: await this.companyService.getAllEmployees(objId) };
+    //   } catch (Error) {
+    //     throw new HttpException(
+    //       'Something went wrong',
+    //       HttpStatus.INTERNAL_SERVER_ERROR,
+    //     );
+    //   }
+    // } else if (
+    //   currentEmployee.role.permissionSuite.includes('view employees under me')
+    // ) {
+    //   validateObjectId(cid);
+    //   try {
+    //     return {
+    //       data: await this.companyService.getAllEmployeeUnderMe(
+    //         body.currentEmployeeId,
+    //       ),
+    //     };
+    //   } catch (Error) {
+    //     throw new HttpException(
+    //       'Something went wrong',
+    //       HttpStatus.INTERNAL_SERVER_ERROR,
+    //     );
+    //   }
+    // } else {
+    //   throw new HttpException('Unauthorized', HttpStatus.UNAUTHORIZED);
+    // }
+    //TODO: Please remove since this is a duplicate of the endpoint that exists in employee.controller.ts
     validateObjectId(cid);
     const objId = new Types.ObjectId(cid);
     try {
       return { data: await this.companyService.getAllEmployees(objId) };
     } catch (Error) {
+      throw new HttpException('Something went wrong', HttpStatus.INTERNAL_SERVER_ERROR);
       throw new HttpException('Something went wrong', HttpStatus.INTERNAL_SERVER_ERROR);
     }
   }
@@ -242,18 +297,31 @@ export class CompanyController {
   })
   @ApiBody({ type: UpdateCompanyDto })
   @Patch('update/:cid')
-  async update(@Headers() headers: any, @Param('cid') cid: string, @Body() updateCompanyDto: UpdateCompanyDto) {
-    try {
-      validateObjectId(cid);
-      const userId = extractUserId(this.jwtService, headers);
+  async update(
+    @Headers() headers: any,
+    @Param('cid') cid: string,
+    @Body()
+    body: {
+      currentEmployeeId: Types.ObjectId;
+      updateCompanyDto: UpdateCompanyDto;
+    },
+  ) {
+    const currentEmployee = await this.employeeService.findById(body.currentEmployeeId);
+    if (currentEmployee.role.permissionSuite.includes('company settings')) {
+      try {
+        validateObjectId(cid);
+        const userId = extractUserId(this.jwtService, headers);
 
-      const companyId = new Types.ObjectId(cid);
-      const updatedCompany = await this.companyService.update(userId, companyId, updateCompanyDto);
-      return {
-        data: updatedCompany,
-      };
-    } catch (e) {
-      throw new HttpException('internal server error', HttpStatus.INTERNAL_SERVER_ERROR);
+        const companyId = new Types.ObjectId(cid);
+        const updatedCompany = await this.companyService.update(userId, companyId, body.updateCompanyDto);
+        return {
+          data: updatedCompany,
+        };
+      } catch (e) {
+        throw new HttpException('internal server error', HttpStatus.INTERNAL_SERVER_ERROR);
+      }
+    } else {
+      throw new HttpException('Unauthorized', HttpStatus.UNAUTHORIZED);
     }
   }
 
@@ -274,15 +342,21 @@ export class CompanyController {
     @Headers() headers: any,
     @Param('cid') companyId: string,
     @UploadedFile() file: Express.Multer.File,
+    @Body() body: { currentEmployeeId: Types.ObjectId },
   ) {
-    try {
-      validateObjectId(companyId);
-      const userId = extractUserId(this.jwtService, headers);
-      return {
-        data: await this.companyService.updateLogo(userId, new Types.ObjectId(companyId), file),
-      };
-    } catch (e) {
-      throw new HttpException('internal server error', HttpStatus.INTERNAL_SERVER_ERROR);
+    const currentEmployee = await this.employeeService.findById(body.currentEmployeeId);
+    if (currentEmployee.role.permissionSuite.includes('company settings')) {
+      try {
+        validateObjectId(companyId);
+        const userId = extractUserId(this.jwtService, headers);
+        return {
+          data: await this.companyService.updateLogo(userId, new Types.ObjectId(companyId), file),
+        };
+      } catch (e) {
+        throw new HttpException('internal server error', HttpStatus.INTERNAL_SERVER_ERROR);
+      }
+    } else {
+      throw new HttpException('Unauthorized', HttpStatus.UNAUTHORIZED);
     }
   }
 
@@ -303,15 +377,24 @@ export class CompanyController {
     description: `The _id attribute of the ${className}`,
   })
   @Delete(':cid')
-  async remove(@Headers() headers: any, @Param('cid') cid: string) {
-    try {
-      validateObjectId(cid);
-      const userId = extractUserId(this.jwtService, headers);
-      const objectId = new Types.ObjectId(cid);
-      await this.companyService.deleteCompany(userId, objectId);
-      return { data: true };
-    } catch (e) {
-      throw new HttpException('Internal Server Error', HttpStatus.SERVICE_UNAVAILABLE);
+  async remove(
+    @Headers() headers: any,
+    @Param('cid') cid: string,
+    @Body() body: { currentEmployeeId: Types.ObjectId },
+  ) {
+    const currentEmployee = await this.employeeService.findById(body.currentEmployeeId);
+    if (currentEmployee.role.permissionSuite.includes('company settings')) {
+      try {
+        validateObjectId(cid);
+        const userId = extractUserId(this.jwtService, headers);
+        const objectId = new Types.ObjectId(cid);
+        await this.companyService.deleteCompany(userId, objectId);
+        return { data: true };
+      } catch (e) {
+        throw new HttpException('Internal Server Error', HttpStatus.SERVICE_UNAVAILABLE);
+      }
+    } else {
+      throw new HttpException('Unauthorized', HttpStatus.UNAUTHORIZED);
     }
   }
 
@@ -332,6 +415,7 @@ export class CompanyController {
   })
   @Delete('/emp')
   async removeEmployee(@Headers() headers: any, @Body() deleteEmployeeDto: DeleteEmployeeFromCompanyDto) {
+    // //TODO: potentially remove the endpoint since there is one in employee.controller.ts
     try {
       const userId = extractUserId(this.jwtService, headers);
       await this.companyService.deleteEmployee(userId, deleteEmployeeDto);
@@ -377,19 +461,31 @@ export class CompanyController {
   })
   @ApiBody({ type: UpdateCompanyJobStatusesDto })
   @Patch('statuses')
-  async updateStatusOrder(@Headers() headers: any, @Body() updateCompanyJobStatusesDto: UpdateCompanyJobStatusesDto) {
-    try {
-      const userId = extractUserId(this.jwtService, headers);
-      const statusArr = new UpdateCompanyJobStatuses(updateCompanyJobStatusesDto);
-      return {
-        data: await this.companyService.updateCompanyStatuses(
-          userId,
-          updateCompanyJobStatusesDto.employeeId,
-          statusArr,
-        ),
-      };
-    } catch (e) {
-      throw e;
+  async updateStatusOrder(
+    @Headers() headers: any,
+    @Body()
+    body: {
+      currentEmployeeId: Types.ObjectId;
+      updateCompanyJobStatusesDto: UpdateCompanyJobStatusesDto;
+    },
+  ) {
+    const currentEmployee = await this.employeeService.findById(body.currentEmployeeId);
+    if (currentEmployee.role.permissionSuite.includes('company settings')) {
+      try {
+        const userId = extractUserId(this.jwtService, headers);
+        const statusArr = new UpdateCompanyJobStatuses(body.updateCompanyJobStatusesDto);
+        return {
+          data: await this.companyService.updateCompanyStatuses(
+            userId,
+            body.updateCompanyJobStatusesDto.employeeId,
+            statusArr,
+          ),
+        };
+      } catch (e) {
+        throw e;
+      }
+    } else {
+      throw new HttpException('Unauthorized', HttpStatus.UNAUTHORIZED);
     }
   }
 }
