+++ src/api-gateway/src/inventory/inventory.controller.ts
@@ -9,6 +9,7 @@ import {
   HttpStatus,
   Post,
   UseGuards,
+  Headers,
 } from '@nestjs/common';
 import { InventoryService } from './inventory.service';
 import { UpdateInventoryDto } from './dto/update-inventory.dto';
@@ -27,13 +28,20 @@ import { Types } from 'mongoose';
 import { BooleanResponseDto } from '../shared/dtos/api-response.dto';
 import { CreateInventoryDto, CreateInventoryResponseDto } from './dto/create-inventory.dto';
 import { AuthGuard } from '../auth/auth.guard';
+// import { extractUserId } from '../utils/Utils';
+import { JwtService } from '@nestjs/jwt';
+import { EmployeeService } from '../employee/employee.service';
 
 const className = 'Inventory';
 
 @ApiTags('Inventory')
 @Controller('inventory')
 export class InventoryController {
-  constructor(private readonly inventoryService: InventoryService) {}
+  constructor(
+    private readonly inventoryService: InventoryService,
+    private readonly jwtService: JwtService,
+    private readonly employeeService: EmployeeService,
+  ) {}
 
   @ApiOperation({
     summary: `Refer to Documentation`,
@@ -60,35 +68,35 @@ export class InventoryController {
     type: CreateInventoryResponseDto,
   })
   @Post('/create')
-  async create(@Body() createInventoryDto: CreateInventoryDto) {
-    let data;
-    try {
-      data = await this.inventoryService.create(createInventoryDto);
-    } catch (e) {
-      throw new HttpException('Invalid request', HttpStatus.BAD_REQUEST);
+  async create(
+    @Headers() headers: any,
+    @Body()
+    body: {
+      currentEmployeeId: Types.ObjectId;
+      createInventoryDto: CreateInventoryDto;
+    },
+  ) {
+    const currentEmployee = await this.employeeService.findById(body.currentEmployeeId);
+    if (currentEmployee.role.permissionSuite.includes('add new inventory item')) {
+      let data;
+      try {
+        data = await this.inventoryService.create(body.createInventoryDto);
+      } catch (e) {
+        throw new HttpException('Invalid request', HttpStatus.BAD_REQUEST);
+      }
+      return { data: data };
+    } else {
+      throw new HttpException('Invalid permission', HttpStatus.BAD_REQUEST);
     }
-    return { data: data };
   }
 
-  @UseGuards(AuthGuard)
-  @ApiBearerAuth('JWT')
-  @ApiInternalServerErrorResponse({
-    type: HttpException,
-    status: HttpStatus.NO_CONTENT,
-  })
-  @ApiOperation({
-    summary: `Get all ${className} instances`,
-    description: `Returns all ${className} instances in the database.`,
-  })
-  @ApiOkResponse({
-    type: InventoryListResponseDto,
-    description: `An array of mongodb objects of the ${className} class.`,
-  })
+  //********Endpoints for test purposes - Start**********/
   @Get('/all')
   async findAll() {
     const data = await await this.inventoryService.findAll();
     return { data: data };
   }
+  //********Endpoints for test purposes - End**********/
 
   @UseGuards(AuthGuard)
   @ApiBearerAuth('JWT')
@@ -113,14 +121,28 @@ export class InventoryController {
     description: `The _id attribute of the Company for which to get all ${className} instances.`,
   })
   @Get('/all/:id')
-  async findAllInCompany(@Param('id') id: Types.ObjectId) {
-    let data;
-    try {
-      data = await this.inventoryService.findAllInCompany(id);
-    } catch (e) {
-      throw new HttpException('Invalid request', HttpStatus.BAD_REQUEST);
+  async findAllInCompany(
+    @Headers() headers: any,
+    @Param('id') id: Types.ObjectId,
+    @Body() body: { currentEmployeeId: Types.ObjectId },
+  ) {
+    console.log('In findAllInCompany');
+    console.log('id', id);
+    const currentEmployee = await this.employeeService.findById(body.currentEmployeeId);
+    console.log('currentEmployee', currentEmployee);
+    if (currentEmployee.role.permissionSuite.includes('view all inventory')) {
+      console.log('in if');
+      let data;
+      try {
+        console.log('in try');
+        data = await this.inventoryService.findAllInCompany(id);
+      } catch (e) {
+        throw new HttpException('Invalid request', HttpStatus.BAD_REQUEST);
+      }
+      return { data: data };
+    } else {
+      throw new HttpException('Invalid permission', HttpStatus.BAD_REQUEST);
     }
-    return { data: data };
   }
 
   @UseGuards(AuthGuard)
@@ -142,9 +164,18 @@ export class InventoryController {
     description: `The _id attribute of the ${className} to be retrieved.`,
   })
   @Get('id/:id')
-  async findById(@Param('id') id: Types.ObjectId) {
-    const data = await this.inventoryService.findById(id);
-    return { data: data };
+  async findById(
+    @Headers() headers: any,
+    @Param('id') id: Types.ObjectId,
+    @Body() body: { currentEmployeeId: Types.ObjectId },
+  ) {
+    const currentEmployee = await this.employeeService.findById(body.currentEmployeeId);
+    if (currentEmployee.role.permissionSuite.includes('view all inventory')) {
+      const data = await this.inventoryService.findById(id);
+      return { data: data };
+    } else {
+      throw new HttpException('Invalid permission', HttpStatus.BAD_REQUEST);
+    }
   }
 
   @UseGuards(AuthGuard)
@@ -167,14 +198,27 @@ export class InventoryController {
   })
   @ApiBody({ type: UpdateInventoryDto })
   @Patch(':id')
-  async update(@Param('id') id: Types.ObjectId, @Body() updateInventoryDto: UpdateInventoryDto) {
-    let data;
-    try {
-      data = await this.inventoryService.update(id, updateInventoryDto);
-    } catch (e) {
-      throw new HttpException('Invalid request', HttpStatus.BAD_REQUEST);
+  async update(
+    @Headers() headers: any,
+    @Param('id') id: Types.ObjectId,
+    @Body()
+    body: {
+      currentEmployeeId: Types.ObjectId;
+      updateInventoryDto: UpdateInventoryDto;
+    },
+  ) {
+    const currentEmployee = await this.employeeService.findById(body.currentEmployeeId);
+    if (currentEmployee.role.permissionSuite.includes('edit all inventory')) {
+      let data;
+      try {
+        data = await this.inventoryService.update(id, body.updateInventoryDto);
+      } catch (e) {
+        throw new HttpException('Invalid request', HttpStatus.BAD_REQUEST);
+      }
+      return { data: data };
+    } else {
+      throw new HttpException('Invalid permission', HttpStatus.BAD_REQUEST);
     }
-    return { data: data };
   }
 
   @UseGuards(AuthGuard)
@@ -201,17 +245,26 @@ export class InventoryController {
     description: `The _id attribute of the ${className}`,
   })
   @Delete(':id')
-  async remove(@Param('id') id: Types.ObjectId) {
-    let data;
-    try {
-      data = await this.inventoryService.remove(id);
-    } catch (e) {
-      throw new HttpException('Invalid request', HttpStatus.BAD_REQUEST);
-    }
+  async remove(
+    @Headers() headers: any,
+    @Param('id') id: Types.ObjectId,
+    @Body() body: { currentEmployeeId: Types.ObjectId },
+  ) {
+    const currentEmployee = await this.employeeService.findById(body.currentEmployeeId);
+    if (currentEmployee.role.permissionSuite.includes('delete inventory item')) {
+      let data;
+      try {
+        data = await this.inventoryService.remove(id);
+      } catch (e) {
+        throw new HttpException('Invalid request', HttpStatus.BAD_REQUEST);
+      }
 
-    if (data === false) {
-      throw new HttpException('update unsuccessful', HttpStatus.INTERNAL_SERVER_ERROR);
+      if (data === false) {
+        throw new HttpException('update unsuccessful', HttpStatus.INTERNAL_SERVER_ERROR);
+      }
+      return { data: data };
+    } else {
+      throw new HttpException('Invalid permission', HttpStatus.BAD_REQUEST);
     }
-    return { data: data };
   }
 }
