+++ src/api-gateway/src/role/role.controller.ts
@@ -9,10 +9,11 @@ import {
   HttpException,
   HttpStatus,
   UseGuards,
+  Headers,
 } from '@nestjs/common';
 import { RoleService } from './role.service';
 import { CreateRoleDto, createRoleResponseDto } from './dto/create-role.dto';
-import { UpdateRoleDto } from './dto/update-role.dto';
+import { UpdateRoleDto, BulkUpdateRoleDto, updateRoleResponseDto } from './dto/update-role.dto';
 import {
   ApiBearerAuth,
   ApiBody,
@@ -27,13 +28,20 @@ import { Types } from 'mongoose';
 import { BooleanResponseDto } from '../shared/dtos/api-response.dto';
 import { RoleListResponseDto, RoleResponseDto } from './entity/role.entity';
 import { AuthGuard } from '../auth/auth.guard';
+import { extractUserId } from '../utils/Utils';
+import { JwtService } from '@nestjs/jwt';
+import { EmployeeService } from '../employee/employee.service';
 
 const className = 'Role';
 
 @ApiTags('Role')
 @Controller('role')
 export class RoleController {
-  constructor(private readonly roleService: RoleService) {}
+  constructor(
+    private readonly roleService: RoleService,
+    private readonly jwtService: JwtService,
+    private readonly employeeService: EmployeeService,
+  ) {}
 
   @ApiOperation({
     summary: `Refer to Documentation`,
@@ -42,21 +50,7 @@ export class RoleController {
   hello() {
     return { message: 'Refer to /documentation for details on the API' };
   }
-
-  @UseGuards(AuthGuard)
-  @ApiBearerAuth('JWT')
-  @ApiInternalServerErrorResponse({
-    type: HttpException,
-    status: HttpStatus.NO_CONTENT,
-  })
-  @ApiOperation({
-    summary: `Get all ${className} instances`,
-    description: `Returns all ${className} instances in the database.`,
-  })
-  @ApiOkResponse({
-    type: RoleListResponseDto,
-    description: `An array of mongodb objects of the ${className} class.`,
-  })
+  //********Endpoints for test purposes - Start**********/
   @Get('/all')
   async findAll() {
     const data = await this.roleService.findAll();
@@ -66,6 +60,18 @@ export class RoleController {
     return { data: data };
   }
 
+  @Get('/createDefault/:id')
+  async createDefault(@Param('id') id: Types.ObjectId) {
+    let data;
+    try {
+      console.log('id', id);
+      data = await this.roleService.createDefaultRoles(id);
+    } catch (e) {
+      throw new HttpException('Invalid request', HttpStatus.BAD_REQUEST);
+    }
+    return { data: data };
+  }
+  //********Endpoints for test purposes - End**********/
   @UseGuards(AuthGuard)
   @ApiBearerAuth('JWT')
   @ApiInternalServerErrorResponse({
@@ -90,9 +96,49 @@ export class RoleController {
   })
   @Get('/all/:id')
   async findAllInCompany(@Param('id') id: Types.ObjectId) {
+    console.log('In findAllInCompany');
     let data;
     try {
+      console.log('In try clause');
       data = await this.roleService.findAllInCompany(id);
+    } catch (e) {
+      throw new HttpException('Invalid request', HttpStatus.BAD_REQUEST);
+    }
+    if (data.length === 0) {
+      throw new HttpException('No data found', HttpStatus.NO_CONTENT);
+    }
+    return { data: data };
+  }
+
+  @UseGuards(AuthGuard)
+  @ApiBearerAuth('JWT')
+  @ApiInternalServerErrorResponse({
+    type: HttpException,
+    status: HttpStatus.NO_CONTENT,
+  })
+  @ApiInternalServerErrorResponse({
+    type: HttpException,
+    status: HttpStatus.BAD_REQUEST,
+  })
+  @ApiOperation({
+    summary: `Get all ${className} instances for a given company`,
+    description: `Returns all ${className} instances in the database for a given Company.`,
+  })
+  @ApiOkResponse({
+    type: RoleListResponseDto,
+    description: `An array of mongodb objects of the ${className} class for a given Company. It excludes the Owner and Default role since those are not allowed to be edited by anyone`,
+  })
+  @ApiParam({
+    name: 'id',
+    description: `The _id attribute of the Company for which to get all ${className} instances.`,
+  })
+  @Get('/all/editing/:id')
+  async findAllInCompanyForEditing(@Param('id') id: Types.ObjectId) {
+    console.log('In findAllInCompanyForEditing');
+    let data;
+    try {
+      console.log('In try clause');
+      data = await this.roleService.findAllInCompanyForEditing(id);
     } catch (e) {
       throw new HttpException('Invalid request', HttpStatus.BAD_REQUEST);
     }
@@ -122,10 +168,13 @@ export class RoleController {
   })
   @Get('id/:id')
   async findById(@Param('id') id: Types.ObjectId) {
+    console.log('In findById endpoint');
     const data = await this.roleService.findById(id);
+    console.log('data: ', data);
     if (!data) {
       throw new HttpException('No data found', HttpStatus.NO_CONTENT);
     }
+    console.log('No exception thrown');
     return { data: data };
   }
 
@@ -161,14 +210,26 @@ export class RoleController {
     type: createRoleResponseDto,
   })
   @Post('/create')
-  async create(@Body() createRoleDto: CreateRoleDto) {
-    let data;
-    try {
-      data = await this.roleService.create(createRoleDto);
-    } catch (e) {
-      throw new HttpException('Invalid request', HttpStatus.BAD_REQUEST);
+  async create(
+    @Headers() headers: any,
+    @Body()
+    body: { currentEmployeeId: Types.ObjectId; createRoleDto: CreateRoleDto },
+  ) {
+    console.log('In create endpoint');
+    const currentEmployee = await this.employeeService.findById(body.currentEmployeeId);
+    console.log('currentEmployee: ', currentEmployee);
+    if (currentEmployee.role.permissionSuite.includes('company settings')) {
+      console.log('In if');
+      let data;
+      try {
+        data = await this.roleService.create(body.createRoleDto);
+      } catch (e) {
+        throw new HttpException('Invalid request', HttpStatus.BAD_REQUEST);
+      }
+      return { data: data };
+    } else {
+      throw new HttpException('Invalid permission', HttpStatus.BAD_REQUEST);
     }
-    return { data: data };
   }
 
   @UseGuards(AuthGuard)
@@ -182,7 +243,7 @@ export class RoleController {
     description: `Send the ${className} ObjectId, and the updated object, and then they get updated if the id is valid.`,
   })
   @ApiOkResponse({
-    type: RoleResponseDto,
+    type: updateRoleResponseDto,
     description: `The updated ${className} object`,
   })
   @ApiParam({
@@ -191,14 +252,71 @@ export class RoleController {
   })
   @ApiBody({ type: UpdateRoleDto })
   @Patch(':id')
-  async update(@Param('id') id: Types.ObjectId, @Body() updateRoleDto: UpdateRoleDto) {
-    let data;
-    try {
-      data = await this.roleService.update(id, updateRoleDto);
-    } catch (e) {
-      throw new HttpException('Invalid request', HttpStatus.BAD_REQUEST);
+  async update(
+    @Headers() headers: any,
+    @Param('id') id: Types.ObjectId,
+    @Body()
+    body: { currentEmployeeId: Types.ObjectId; updateRoleDto: UpdateRoleDto },
+  ) {
+    console.log('In update endpoint');
+    const userId = extractUserId(this.jwtService, headers);
+    const currentEmployee = await this.employeeService.findById(body.currentEmployeeId);
+    console.log('currentEmployee: ', currentEmployee);
+    if (currentEmployee.role.permissionSuite.includes('company settings')) {
+      let data;
+      try {
+        data = await this.roleService.update(userId, id, body.updateRoleDto);
+      } catch (e) {
+        throw new HttpException('Invalid request', HttpStatus.BAD_REQUEST);
+      }
+      return { data: data };
+    } else {
+      throw new HttpException('Invalid permission', HttpStatus.BAD_REQUEST);
     }
-    return { data: data };
+  }
+
+  @UseGuards(AuthGuard)
+  @ApiBearerAuth('JWT')
+  @ApiInternalServerErrorResponse({
+    type: HttpException,
+    status: HttpStatus.BAD_REQUEST,
+  })
+  @ApiOperation({
+    summary: `Update an ${className} instances`,
+    description: `Send array of ${className} ObjectIds and an array of updated objects, and then they get updated if the id is valid.`,
+  })
+  @ApiOkResponse({
+    type: [updateRoleResponseDto],
+    description: `The updated ${className} object`,
+  })
+  @ApiBody({ type: BulkUpdateRoleDto })
+  @Patch('/bulkUpdate/:companyId')
+  async bulkUpdate(
+    @Headers() headers: any,
+    @Param('companyId') companyId: Types.ObjectId,
+    @Body()
+    body: {
+      currentEmployeeId: Types.ObjectId;
+      updateRoleDto: BulkUpdateRoleDto;
+    },
+  ) {
+    console.log('In bulkUpdate endpoint');
+    const userId = extractUserId(this.jwtService, headers);
+    const currentEmployee = await this.employeeService.findById(body.currentEmployeeId);
+    console.log('currentEmployee: ', currentEmployee);
+    if (currentEmployee.role.permissionSuite.includes('company settings')) {
+      let data;
+      console.log('In if');
+      try {
+        console.log('In try clause');
+        data = await this.roleService.bulkUpdate(userId, body.updateRoleDto, companyId);
+      } catch (e) {
+        throw new HttpException('Invalid request', HttpStatus.BAD_REQUEST);
+      }
+      return { data: data };
+    } else {
+      throw new HttpException('Invalid permission', HttpStatus.BAD_REQUEST);
+    }
   }
 
   @UseGuards(AuthGuard)
@@ -225,17 +343,26 @@ export class RoleController {
     description: `The _id attribute of the ${className}`,
   })
   @Delete(':id')
-  async remove(@Param('id') id: Types.ObjectId) {
-    let data;
-    try {
-      data = await this.roleService.remove(id);
-    } catch (e) {
-      throw new HttpException('Invalid request', HttpStatus.BAD_REQUEST);
-    }
+  async remove(
+    @Headers() headers: any,
+    @Param('id') id: Types.ObjectId,
+    @Body() body: { currentEmployeeId: Types.ObjectId },
+  ) {
+    const currentEmployee = await this.employeeService.findById(body.currentEmployeeId);
+    if (currentEmployee.role.permissionSuite.includes('company settings')) {
+      let data;
+      try {
+        data = await this.roleService.remove(id);
+      } catch (e) {
+        throw new HttpException('Invalid request', HttpStatus.BAD_REQUEST);
+      }
 
-    if (data === false) {
-      throw new HttpException('update unsuccessful', HttpStatus.INTERNAL_SERVER_ERROR);
+      if (data === false) {
+        throw new HttpException('update unsuccessful', HttpStatus.INTERNAL_SERVER_ERROR);
+      }
+      return { data: data };
+    } else {
+      throw new HttpException('Invalid permission', HttpStatus.BAD_REQUEST);
     }
-    return { data: data };
   }
 }
